Nit clean up
