+++ apps/studio/components/interfaces/Support/AttachmentUpload.tsx
@@ -1,3 +1,4 @@
+import { createClient, type SupabaseClient } from '@supabase/supabase-js'
 import { compact } from 'lodash'
 import { Plus, X } from 'lucide-react'
 import { type ChangeEvent, useCallback, useEffect, useMemo, useRef, useState } from 'react'
@@ -6,11 +7,30 @@ import { toast } from 'sonner'
 
 import { uuidv4 } from 'lib/helpers'
 import { cn } from 'ui'
-import { createSupportStorageClient } from './support-storage-client'
 
 const MAX_ATTACHMENTS = 5
 
-export const uploadAttachments = async (ref: string, files: File[]) => {
+export const createSupportStorageClient = (): SupabaseClient => {
+  const SUPPORT_API_URL = process.env.NEXT_PUBLIC_SUPPORT_API_URL || ''
+  const SUPPORT_API_KEY = process.env.NEXT_PUBLIC_SUPPORT_ANON_KEY || ''
+
+  return createClient(SUPPORT_API_URL, SUPPORT_API_KEY, {
+    auth: {
+      persistSession: false,
+      autoRefreshToken: false,
+      // @ts-expect-error
+      multiTab: false,
+      detectSessionInUrl: false,
+      localStorage: {
+        getItem: (_key: string) => undefined,
+        setItem: (_key: string, _value: string) => {},
+        removeItem: (_key: string) => {},
+      },
+    },
+  })
+}
+
+const uploadAttachments = async (ref: string, files: File[]) => {
   const supportSupabaseClient = createSupportStorageClient()
 
   const filesToUpload = Array.from(files)
